flask-app/pyproject.toml
[tool.poetry]
name = "flask-app"
version = "0.1.0"
description = ""
authors = ["bhattjaldhi <bhattjaldhi27@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
python-dotenv = "^1.0.1"
flask-socketio = "^5.3.7"
duckdb = "^1.1.1"
flask-pydantic = "^0.12.0"
pytest = "^8.3.3"
pytest-flask = "^1.3.0"
pytest-cov = "^5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
